feat: Separate InputArea, LinkArea components
